feat(plugins): add the ability to disable todo-comments.nvim in zen-mode (#114)
I'm an avid and happy user of the
[todo-comments.nvim](https://github.com/folke/todo-comments.nvim)
plugin, but when writing (text, not code) in zen mode, I personally find
the highlights to be distracting sometimes. Since
[todo-comments.nvim](https://github.com/folke/todo-comments.nvim) is
also part of the @folke ecosystem of plugins, I thought an option to
explicitly disable the highlights made by the plugin would make a nice
addition to the zen-mode.nvim.