fix: tmux+wezterm escape sequences
This fixes two problems with the previous implementation,
first: use the `-w0` argument to the `base64` command to
disable line wrapping so it doesn't output a new line
character.

second: NeoVim seems to clear escape sequences before
sending it into stdout, the only workaround I cound find
was to use the call command to the chansend function.

Relevant links:
https://github.com/neovim/neovim/issues/1496
https://github.com/neovim/neovim/issues/2765